pipeline/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
feedparser==6.0.11
psycopg[binary]==3.1.18
python-dotenv==1.0.1
pytest
langchain
openai
python-dotenv
langchain_community
bs4
python-dateutil
